Refactor entry point
